feat(): engagement Score implemented

+++ src/main/java/dev/pantanal/b3/krpv/acao_social/modulos/person/dto/response/PersonResponseDto.java
@@ -13,7 +13,7 @@ public record PersonResponseDto (
         LocalDateTime dateBirth,
         StatusEnum status,
         String cpf,
-        BigInteger engagementScore,
+        Integer engagementScore,
         UUID createdBy,
         LocalDateTime createdDate,
         UUID lastModifiedBy,
